# refactor(analyzer): use packages.Load instead of build.Import to get better performance