… go.sum
… vendor/github.com/fsnotify/fsnotify/.editorconfig
… vendor/github.com/fsnotify/fsnotify/.gitattributes
… vendor/github.com/fsnotify/fsnotify/.travis.yml
… vendor/github.com/fsnotify/fsnotify/LICENSE
… vendor/github.com/fsnotify/fsnotify/README.md
… vendor/github.com/fsnotify/fsnotify/fsnotify.go
… vendor/github.com/fsnotify/fsnotify/go.mod
… vendor/github.com/fsnotify/fsnotify/go.sum
… vendor/github.com/fsnotify/fsnotify/inotify_poller.go
… vendor/github.com/fsnotify/fsnotify/open_mode_bsd.go
… vendor/github.com/fsnotify/fsnotify/open_mode_darwin.go
… vendor/github.com/go-openapi/jsonreference/go.mod
… vendor/github.com/go-openapi/jsonreference/go.sum
… vendor/github.com/golang/protobuf/proto/buffer.go
… vendor/github.com/golang/protobuf/proto/clone.go
… vendor/github.com/golang/protobuf/proto/decode.go
… vendor/github.com/golang/protobuf/proto/defaults.go
… vendor/github.com/golang/protobuf/proto/deprecated.go
… vendor/github.com/golang/protobuf/proto/discard.go
… vendor/github.com/golang/protobuf/proto/encode.go
… vendor/github.com/golang/protobuf/proto/equal.go
… vendor/github.com/golang/protobuf/proto/extensions.go
… vendor/github.com/golang/protobuf/proto/lib.go
… vendor/github.com/golang/protobuf/proto/message_set.go
… vendor/github.com/golang/protobuf/proto/pointer_reflect.go
… vendor/github.com/golang/protobuf/proto/pointer_unsafe.go
… vendor/github.com/golang/protobuf/proto/properties.go
… vendor/github.com/golang/protobuf/proto/proto.go
… vendor/github.com/golang/protobuf/proto/registry.go