<!--  Thanks for sending a pull request! See below for tips! -->

**What this PR does / why we need it**:

There are basically two changes: first, use `packages.Load` instead of `build.Import` to analyze a package, and second, import all packages at once rather than individually.

Take storage-admin project as an example, API document generation:

<table>
<thead><tr><th>Before</th><th>After</th></tr></thead>
<tbody>

<tr><td>

```
1110.53s user
1321.23s system
499% cpu
8:06.60 total
```

</td>

<td>

```
144.90s user
23.57s system
610% cpu
27.607 total
```

</td></tr>
</tbody></table>



API client generation:

<table>
<thead><tr><th>Before</th><th>After</th></tr></thead>
<tbody>

<tr><td>

```
1070.64s user
1265.10s system
501% cpu
7:45.53 total
```

</td>

<td>

```
138.41s user
22.62s system
626% cpu
25.691 total
```

</td></tr>
</tbody></table>

The average time reduced from 8 minutes to half a minute.

**Which issue(s) this PR is related to** *(optional, link to 3rd issue(s))*:

Fixes #

Reference to #
<!-- 填在 Fixes，PR 合并就会关 issue。填在 Reference to 会关联 issue，不会联动关闭。-->

**Special notes for your reviewer**:

/cc @your-reviewer

<!-- Please answer the following questions during the code freeze, and delete this line.
**Code freeze questions**

1. What causes this PR to not be merged before code freeze?
2. Why this PR is absolutely necessary for this version? Paste a screenshot of smoke testing docs if you could.
3. What's the effects after merging it?
4. Is there anyway we can skip this to not affect the overall process?
-->

/kind feature

**Release note**:
<!--  Write your release note:
1. Enter your extended release note in the below block. If the PR requires additional action from users switching to the new release, include the string "action required".
2. If no release note is required, just write "NONE".
-->

```release-note
Use `packages.Load` instead of `build.Import` in API documentation/client generation to get better performance.
```

<!--  Thanks for sending a pull request! Here are some tips:

1. https://github.com/caicloud/engineering/blob/master/guidelines/review_conventions.md      <-- what is the review process looks like
2. https://github.com/caicloud/engineering/blob/master/guidelines/git_commit_conventions.md  <-- how to structure your git commit
3. https://github.com/caicloud/engineering/blob/master/guidelines/caicloud_bot.md            <-- how to work with caicloud bot

Other tips:

If this is your first contribution, read our Getting Started guide https://github.com/caicloud/engineering/blob/master/guidelines/README.md
-->